Make plunker integrated
