+++ src/app/app-module.ts
@@ -6,8 +6,6 @@ import {HttpModule} from '@angular/http';
 import {MaterialModule} from '@angular/material';
 import {MaterialDocsApp} from './material-docs-app';
 import {Homepage} from './pages/homepage/homepage';
-import {NavBar} from './shared/navbar/navbar';
-import {PlunkerButton} from './shared/plunker';
 import {routing} from './routes';
 import {ComponentList} from './pages/component-list/component-list';
 import {ComponentViewer} from './pages/component-viewer/component-viewer';
@@ -21,8 +19,6 @@ import {SharedModule} from './shared/shared-module';
     ComponentList,
     ComponentViewer,
     Homepage,
-    NavBar,
-    PlunkerButton,
   ],
   imports: [
     BrowserModule,
