--- src/app/shared/plunker/example-data.ts
@@ -1,17 +1,0 @@
-/**
- * Example data
- *   with information about Component name, selector, files used in example, and path to examples
- *   TODO(tinayuangao): find a better place for this file
- */
-export class ExampleData {
-  // TODO: figure out how do we get these variables.
-  description: string = 'Some description for material';
-  // TODO: use real example and delete the example/ folder.
-  examplePath = '/assets/example/';
-  exampleFiles = ['button-demo.html', 'button-demo.scss', 'button-demo.ts'];
-
-  // TODO: extract these variables from example code.
-  selectorName = 'button-demo';
-  indexFilename = 'button-demo';
-  componentName = 'ButtonDemo';
-}
