+++ src/app/shared/plunker/plunker-button.html
@@ -1,2 +1,2 @@
 <!-- TODO: change the template to be plunker icon -->
-<button md-button (click)="openPlunker()">Open Plunker</button>
+<button md-icon-button type="button"(click)="openPlunker()">P</button>
